fix: add suggested changes
