+++ App.tsx
@@ -1,23 +1,13 @@
-import {useMemo, useState} from 'react';
+import {useState} from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 
 import SplashScreen from '@screens/splash';
-import Login from '@screens/login';
 
-import MainNavigator from '@navigation/MainNavigator';
 import {ThemeContextProvider} from '@config/contexts/ThemeContext';
+import MainNavigator from '@config/MainNavigator';
 
 function App() {
   const [showSplashScreen, setShowSplashScreen] = useState(true);
-  const [isUserLoggedIn, setIsUserLoggedIn] = useState(false);
-
-  const navigator = useMemo(() => {
-    return isUserLoggedIn ? (
-      <MainNavigator />
-    ) : (
-      <Login makeUserLogin={() => setIsUserLoggedIn(true)} />
-    );
-  }, [isUserLoggedIn, setIsUserLoggedIn]);
 
   return (
     <ThemeContextProvider>
@@ -25,7 +15,7 @@ function App() {
         {showSplashScreen ? (
           <SplashScreen hideSplashScreen={() => setShowSplashScreen(false)} />
         ) : (
-          navigator
+          <MainNavigator />
         )}
       </NavigationContainer>
     </ThemeContextProvider>
